calc_prod_igr

This function calculates production using the instantaneous growth rate method.

Usage

calc_prod_igr(
  taxaSampleListMass = NULL,
  taxaInfo = NULL,
  bootNum = NULL,
  dateDf = NULL,
  taxaSummary = "full",
  wrap = TRUE,
  massValue = NULL,
  massLabel = NULL,
  ...
)

Arguments

taxaSampleListMass

description

taxaInfo

The taxa info data.frame

bootNum

integer. The number of bootstrap samples to produce.

dateDf

data frame of date information with external predictors for each month. There should be a column name identical to all variables in the growth equation found in taxaInfo data.frame.

taxaSummary

string of 'short', 'full', or 'none'. What type of summary information should be returned.

wrap

logical. Should the dates wrap to create a full year?

massValue

string. What is the mass value and units of the production

massLabel

string. What label should the output units be. It is possible this will default to 'massValue' in the future.

...

additional arguments to pass to the function

Value

returns a list of 2 objects

Details

Additional details...