calc_prod_sf

This function calculates secondary production with the size-frequency method.

Usage

calc_prod_sf(
  taxaSampleListMass = NULL,
  taxaInfo = NULL,
  bootNum = NULL,
  dateDf = NULL,
  taxaSummary = "full",
  wrap = FALSE,
  lengthValue = NULL,
  massValue = "afdm_mg",
  abunValue = "density",
  dateCol = "dateID",
  repCol = "repID",
  bootList = NULL,
  ...
)

Arguments

taxaSampleListMass

description

taxaInfo

data frame of taxonomic information for calculating production

bootNum

integer. How many bootstrap samples should be constructed

dateDf

data frame of date information with external predictors for each month. There should be a column name identical to all variables in the growth equation found in taxaInfo data.frame.

taxaSummary

string of 'short', 'full', or 'none'. What type of summary information should be returned.

wrap

logical. Should the dates wrap to create a full year?

massValue

string. What is the mass value and units of the production

bootList

list. This is the bootstrapped samples passed from `calc_production()`

...

additional arguments to be passed to the function

massLabel

string. What label should the output units be. It is possible this will default to 'massValue' in the future.

Value

returns a list of 2 objects:

P.boots: the boostrapped estimates of production, abundance, and biomass.

taxaSummary: is the summary of the sample production, abundance, and biomass